Add ephemeral option to rolenotifier
The `ephemeral` option is available when a channel is nominated as the target for a role notification. If set to `true` (the default is `false`) this will cause the bot to send a message and immediately remove it. 

For this reason, you probably don't need this message to be anything other than {2}, though you can add more information to be delivered via mobile push notifications (deleting the message generally doesn't delete the notification, and so you can provide more context about why the notification was sent).